renaming-texts-and-finish-styling
