+++ src/components/artobject/ArtObject.js
@@ -9,10 +9,14 @@ function ArtObject({artName, artImage, artDescription, artInfo, art, material, n
             <img src={artImage} alt="art-image" width="500px"/>
             <h4>Extra titels: </h4>
             {nameTag && nameTag.map((name, index) => <h4 key={index}>{name}</h4>)}
+            <br/>
             <h4>Omschrijving: {artDescription && artDescription}</h4>
+            <br/>
             <h4>Kunstwerk type(s): </h4>
             {art && art.map((art, index) => <h4 key={index}>{art}</h4>)}
+            <br/>
             <h4>Soort kunst: {artInfo && artInfo.physicalMedium}</h4>
+            <br/>
             <h4>Materialen: </h4>
             {material && material.map((material, index) => <h4 key={index}>{material}</h4>)}
         </div>
