+++ src/components/artpiece/ArtPiece.js
@@ -9,10 +9,11 @@ function ArtPiece({artObjects}) {
                 return <div className="art-container" key={artObject.id}>
                     <ul key={artObject.objectNumber} className="art-body">
                         <h2><Link to={`/objectInfo/${artObject.objectNumber}`}>{artObject.title}</Link></h2>
-                        <p>{artObject.objectNumber}</p>
+                        <p>Museum-code: {artObject.id}</p>
                         {artObject.webImage !== null ?
                             <img src={artObject.webImage.url} alt="art-image" width="300px"/> : ''}
-                        <p>{artObject.id}</p>
+                        <p>kunsternaar(-es):</p>
+                        <p>{artObject.principalOrFirstMaker}</p>
                     </ul>
                 </div>
             })}
