+++ src/pages/contact/Contact.js
@@ -6,15 +6,15 @@ const Contact = () => {
     return (
         <Fragment>
             <h3>Contactpagina:</h3>
-            <p> Dit project is gemaakt m.b.v. Jetbrains Webstorm. </p>
-            <p> Dit is een eindproject gemaakt door Marcel Brzezinski, woonachtig in Horst aan de Maas ((Noord-)Limburg)</p>
-            <p> Ik ben een student aan de Hogeschool Novi te Utrecht, en mijn docenten zijn:</p>
+            <h4> Dit project is gemaakt m.b.v. Jetbrains Webstorm. </h4>
+            <h4> Dit is een eindproject gemaakt door Marcel Brzezinski, woonachtig in Horst aan de Maas ((Noord-)Limburg)</h4>
+            <h4> Ik ben een student aan de Hogeschool Novi te Utrecht, en mijn docenten zijn:</h4>
             <ul>
                 <li>Nova Eeken (Hoofd-docent (Front-end))</li>
                 <li>Sam Barnhoorn (Subject Matter Expert)</li>
             </ul>
-            <p>Dit project dat U voor ziet is het resultaat van een halfjaar bootcamp front-end webdevelopment!</p>
-            <p>Zoals U ziet is een eindproject over het Rijksmuseum te Amsterdam.</p>
+            <h4>Dit project dat U voor ziet is het resultaat van een halfjaar bootcamp front-end webdevelopment!</h4>
+            <h4>Zoals U ziet is een eindproject over het Rijksmuseum te Amsterdam.</h4>
             <footer></footer>
         </Fragment>
