+++ src/pages/home/Home.css
@@ -9,6 +9,10 @@
     float: left;
 }
 
+p {
+    color:black;
+    background-color: white;
+}
 
 ul {
     padding-left: 20px;
